+++ android/app/src/main/kotlin/net/mullvad/mullvadvpn/compose/test/ComposeTestTagConstants.kt
@@ -24,5 +24,6 @@ const val LOCATION_INFO_TEST_TAG = "location_info_test_tag"
 
 // ConnectScreen - Notification banner
 const val NOTIFICATION_BANNER = "notification_banner"
+const val NOTIFICATION_BANNER_ACTION = "notification_banner_action"
 
 const val LOGIN_TITLE_TEST_TAG = "login_title_test_tag"