+++ android/app/src/main/kotlin/net/mullvad/mullvadvpn/ui/fragment/ConnectFragment.kt
@@ -48,7 +48,8 @@ class ConnectFragment : BaseFragment() {
                     onManageAccountClick = connectViewModel::onManageAccountClick,
                     onOpenOutOfTimeScreen = ::openOutOfTimeScreen,
                     onSettingsClick = ::openSettingsView,
-                    onAccountClick = ::openAccountView
+                    onAccountClick = ::openAccountView,
+                    onDismissNewDeviceClick = connectViewModel::dismissNewDeviceNotification,
                 )
             }
         }